requirements.txt
pandas
openpyxl
streamlit
numpy
sympy
math
matplotlib.pyplot
scipy